test: split test files per-predicate
Makes it simpler to run benchmarks as groups with one click per file.
